LP library are now taken care of by plugin starting with lp_lib_
