+++ plugins/lp_lib_cplex/wscript
@@ -1,0 +1,50 @@
+#! /usr/bin/env python
+# encoding: utf-8
+
+import os
+
+######################
+###### options #######
+######################
+def options (opt):
+	grp_name = "cplex options (when --lp-lib=cplex is used)"
+	grp = opt.add_option_group (grp_name)
+	grp.add_option ("--cplex-path", action="store", type="string",
+									dest="CPLEX_PATH", default = "",
+	                help = "location of the cplex lib and include directories \
+	                        (by default use the one in 3rd directory)")
+
+######################
+##### configure ######
+######################
+def configure (conf):
+	if conf.env["LP_LIB"]:
+		conf.fatal ("Trying to configure a second library for LP")
+	conf.env["LP_LIB"] = "CPLEX"
+
+	cplex_dir = conf.options.CPLEX_PATH
+	if cplex_dir:
+		cplex_include = os.path.join (cplex_dir, "include")
+		cplex_lib = os.path.join (cplex_dir, "lib/x86-64_linux/static_pic/")
+		conf.env.append_unique ("INCLUDES_IBEX_DEPS", cplex_include)
+		conf.env.append_unique ("LIBPATH_IBEX_DEPS", cplex_lib)
+	else:
+		cplex_include = ""
+		cplex_lib = ""
+
+	conf.check_cxx (header_name = "ilcplex/cplex.h", includes = cplex_include,
+			use = ["IBEX", "LP_LIB"], uselib_store = "LP_LIB")
+	conf.check_cxx (lib = ["cplex", "pthread"], libpath = cplex_lib,
+			use = ["IBEX", "LP_LIB"], uselib_store = "LP_LIB")
+
+	conf.env.IBEX_LP_LIB_INCLUDES = '#include "ilcplex/cplex.h"'
+	conf.env.IBEX_LP_LIB_EXTRA_ATTRIBUTES = """
+	CPXENVptr  envcplex;
+	CPXLPptr lpcplex;
+	int * indice;
+	double * tmp;
+	int * r_matbeg;
+	double * r_matval;
+	int * r_matind;
+	"""
+	conf.env.IBEX_LP_LIB_DEFAULT_EPS = "1e-9"
