+++ plugins/lp_lib_none/wscript
@@ -1,0 +1,22 @@
+#! /usr/bin/env python
+# encoding: utf-8
+
+import os
+
+######################
+###### options #######
+######################
+def options (opt):
+  pass
+
+######################
+##### configure ######
+######################
+def configure (conf):
+	if conf.env["LP_LIB"]:
+		conf.fatal ("Trying to configure a second library for LP")
+	conf.env["LP_LIB"] = "NONE"
+
+	conf.env.IBEX_LP_LIB_INCLUDES = ""
+	conf.env.IBEX_LP_LIB_EXTRA_ATTRIBUTES = ""
+	conf.env.IBEX_LP_LIB_DEFAULT_EPS = "1e-10"
