+++ src/numeric/ibex_LPException.h
